Separate `Rectangle` from `RectangleF` extensions
