+++ source/MonoGame.Extended/RectangleF.Extensions.cs
@@ -1,0 +1,39 @@
+﻿using Microsoft.Xna.Framework;
+
+namespace MonoGame.Extended
+{
+    public static class RectangleFExtensions
+    {
+        /// <summary>
+        ///     Gets the corners of the rectangle in a clockwise direction starting at the top left.
+        /// </summary>
+        public static Vector2[] GetCorners(this RectangleF rectangle)
+        {
+            var corners = new Vector2[4];
+            corners[0] = new Vector2(rectangle.Left, rectangle.Top);
+            corners[1] = new Vector2(rectangle.Right, rectangle.Top);
+            corners[2] = new Vector2(rectangle.Right, rectangle.Bottom);
+            corners[3] = new Vector2(rectangle.Left, rectangle.Bottom);
+            return corners;
+        }
+
+        public static Rectangle ToRectangle(this RectangleF rectangle)
+        {
+            return new Rectangle((int) rectangle.X, (int) rectangle.Y, (int) rectangle.Width, (int) rectangle.Height);
+        }
+
+        public static RectangleF Clip(this RectangleF rectangle, RectangleF clippingRectangle)
+        {
+            var clip = clippingRectangle;
+            rectangle.X = clip.X > rectangle.X ? clip.X : rectangle.X;
+            rectangle.Y = clip.Y > rectangle.Y ? clip.Y : rectangle.Y;
+            rectangle.Width = rectangle.Right > clip.Right ? clip.Right - rectangle.X : rectangle.Width;
+            rectangle.Height = rectangle.Bottom > clip.Bottom ? clip.Bottom - rectangle.Y : rectangle.Height;
+
+            if(rectangle.Width <= 0 || rectangle.Height <= 0)
+                return RectangleF.Empty;
+
+            return rectangle;
+        }
+    }
+}
